% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analCI.R
\name{analCI}
\alias{analCI}
\title{Calcium Image data analysis.}
\usage{
analCI(grupos = NULL, agrupacion = "silueta", modo = "Kmedioids",
  outlier = TRUE, directory = NULL, skip = 5, data.scale = TRUE,
  legend.ROIs = TRUE, interval = NULL, Units = "ms", Smooth. = TRUE)
}
\arguments{
\item{grupos}{Allows stablishing a number of groups deshired}
}
\value{
plots
}
\description{
this function allows arring out a comprensive calcium image data
  analysis. Notably, the folder which stores the .txt also has to store a
  .csv file where each row correspond to a stimuli, shuch that the first
  colomn store the name of each stimuli, the second column the stimuli start
  time, in seconds, and the third column the stimuli-end-time. Aditional rows
  can be typing: cut: This row meaning the interval of experiment which you
  want to remove. Nisoldipina: The interval of experiment where Nisoldipina,
  an antagonist of VOCCs, is emplyed in order to remove the masking efect of
  Calcium entry through VOCCs when you are concern with asses SOCE
}
\author{
Enrique Perez_Riesgo
}
