% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analCI.R
\name{mahOutlier}
\alias{mahOutlier}
\title{Multivariate outliers.}
\usage{
mahOutlier(X)
}
\value{
plots
}
\description{
An oultier detection is carried out by applying Mahalanobis distances, where means vetor and correlation matrix is computed with those data whose mahalanobis distance is lower than median of mahalanobis distances.
}
\author{
Enrique Perez_Riesgo
}
