% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analCI.R
\name{wave.length}
\alias{wave.length}
\title{wave length}
\usage{
wave.length(interval = NULL, data)
}
\value{
plots
}
\description{
This functions provide the number of peaks, oscilations.
}
\author{
Enrique Perez_Riesgo
}
